pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "configlint"
version = "1.0.0"
description = "🔍 ConfigLint - Universal Configuration File Linter & Auto-Fixer | 通用配置文件语法检查与自动修复工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "gitstq", email = "gitstq@users.noreply.github.com"}
]
keywords = [
    "config", "linter", "validation", "json", "yaml", "toml", "ini", "env",
    "syntax-check", "auto-fix", "cli", "developer-tools"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "jsonschema>=4.0.0",
    "pyyaml>=6.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
configlint = "configlint.cli:main"

[project.urls]
Homepage = "https://github.com/gitstq/ConfigLint"
Repository = "https://github.com/gitstq/ConfigLint"
Issues = "https://github.com/gitstq/ConfigLint/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
configlint = ["py.typed"]
